+++ src/Testcontainers/Configurations/Modules/Databases/MsSqlTestcontainerConfiguration.cs
@@ -8,11 +8,11 @@ namespace DotNet.Testcontainers.Configurations
   [PublicAPI]
   public class MsSqlTestcontainerConfiguration : TestcontainerDatabaseConfiguration
   {
-    internal const string DefaultDatabase = "master";
+    public const string MasterDatabase = "master";
+
     private const string MsSqlImage = "mcr.microsoft.com/mssql/server:2017-CU28-ubuntu-16.04";
 
     private const int MsSqlPort = 1433;
-    private string database = DefaultDatabase;
 
     /// <summary>
     /// Initializes a new instance of the <see cref="MsSqlTestcontainerConfiguration" /> class.
@@ -33,11 +33,8 @@ namespace DotNet.Testcontainers.Configurations
     }
 
     /// <inheritdoc />
-    public override string Database
-    {
-      get => this.database;
-      set => this.database = value;
-    }
+    public override string Database { get; set; }
+      = MasterDatabase;
 
     /// <inheritdoc />
     public override string Username